requirements.txt
fastapi
uvicorn
python-dotenv
livekit-server-sdk
